# app/static/openapi.yaml
openapi: 3.0.3
info:
  title: Verso-Backend API
  description: |
    REST API for the Verso-Backend sovereign business operating system.
    
    ## Authentication
    
    All API endpoints require Bearer token authentication. Include your API key in the Authorization header:
    
    ```
    Authorization: Bearer sk_live_your_api_key_here
    ```
    
    ## Rate Limiting
    
    API requests are rate limited to:
    - 100 requests per minute for read operations
    - 30 requests per minute for write operations
    
    Rate limit headers are included in responses:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Requests remaining in window
    - `X-RateLimit-Reset`: Timestamp when limit resets
    
    ## Pagination
    
    List endpoints support pagination via query parameters:
    - `page`: Page number (default: 1)
    - `per_page`: Items per page (default: 25, max: 100)
    
    Paginated responses include a `pagination` object with total count and page info.
    
    ## Errors
    
    Errors return appropriate HTTP status codes with JSON body:
    ```json
    {
      "error": "Error message description"
    }
    ```
    
  version: 1.0.0
  contact:
    name: Verso Industries
    url: https://versoindustries.com
    email: support@versoindustries.com
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
servers:
  - url: https://your-domain.com/api/v1
    description: Production server
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API key obtained from Admin Dashboard → Manage API Keys
  schemas:
    Lead:
      type: object
      properties:
        id:
          type: integer
          example: 1
        first_name:
          type: string
          example: "John"
        last_name:
          type: string
          example: "Doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        phone:
          type: string
          example: "+1-555-123-4567"
        message:
          type: string
          example: "I'm interested in your services"
        status:
          type: string
          enum: [New, Contacted, Qualified, Proposal, Won, Lost]
          example: "New"
        source:
          type: string
          example: "website"
        notes:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
          example: ["enterprise", "urgent"]
        submitted_at:
          type: string
          format: date-time
          example: "2024-12-01T10:30:00Z"
    LeadCreate:
      type: object
      required:
        - first_name
        - last_name
        - email
        - phone
        - message
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        message:
          type: string
        status:
          type: string
          default: "New"
        source:
          type: string
          default: "api"
        notes:
          type: string
        tags:
          type: array
          items:
            type: string
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 25
        total:
          type: integer
          example: 150
        pages:
          type: integer
          example: 6
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Error message"
paths:
  /leads:
    get:
      tags:
        - Leads
      summary: List all leads
      description: Get all leads with optional filtering, sorting, and pagination.
      operationId: getLeads
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 25
            maximum: 100
        - name: sort
          in: query
          schema:
            type: string
            default: submitted_at
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: status
          in: query
          schema:
            type: string
        - name: created_after
          in: query
          schema:
            type: string
            format: date-time
        - name: created_before
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: List of leads
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Lead'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []

    post:
      tags:
        - Leads
      summary: Create a lead
      description: Create a new lead in the CRM system.
      operationId: createLead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeadCreate'
      responses:
        '201':
          description: Lead created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  message:
                    type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
...